removed pathname from mobile screens
