+++ src/components/pages/classroom/classroom-routes/record-grade/score-board/index.tsx
